move BalanceInfo to types
